requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
fake-useragent==1.5.1
idna==3.8
pydantic==2.9.1
pydantic_core==2.23.3
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
loguru~=0.7.2
SQLAlchemy~=2.0.34
python-dotenv~=1.0.1